refactor: update pulse.js to use new API endpoint and restructure client initialization
