+++ apps/frontend/static/pulse.js
@@ -1,1 +1,1 @@
-var d=async o=>{let t="https://api.pulsed.workers.dev/ingest";if(o.events.length===0)return;let n=JSON.stringify(o);if(navigator.sendBeacon){let e=new Blob([n],{type:"application/json"});if(navigator.sendBeacon(t,e))return}try{await fetch(t,{method:"POST",headers:{"Content-Type":"application/json"},body:n,keepalive:!0})}catch(e){console.error("Analytics: Failed to send batch",e)}};var f={batchSize:10,timeout:5e3},l=o=>{let t={...f,...o},n=[],e=null,r=async()=>{if(e&&(clearTimeout(e),e=null),n.length===0)return;let i={apiKey:t.apiKey,context:t.context,events:[...n]};n=[],await d(i)};return{add:i=>{if(n.push({...i}),n.length>=t.batchSize){r();return}e&&clearTimeout(e),e=setTimeout(r,t.timeout)},flush:r}};var p=()=>{let o="analytics_visitor_id",t=localStorage.getItem(o);return t||(t=m(),localStorage.setItem(o,t)),t},h=()=>{let t=sessionStorage.getItem("_sid"),n=sessionStorage.getItem("_sid_t"),e=Date.now();if(t&&n&&e-parseInt(n)<18e5)return sessionStorage.setItem("_sid_t",e.toString()),t;let r=m();return sessionStorage.setItem("_sid",r),sessionStorage.setItem("_sid_t",e.toString()),r},m=()=>Date.now().toString(36)+Math.random().toString(36).substring(2);var g=(o,t={},{batcher:n,visitorId:e,sessionId:r})=>{let a={type:o,visitorId:e,sessionId:r,timestamp:Date.now(),properties:{...t,path:window.location.pathname,query:window.location.search,title:document.title,...o==="page_view"&&{referrer:document.referrer}}};n.add(a)},u=({batcher:o,visitorId:t,sessionId:n})=>{document.addEventListener("click",e=>{let a=e.target.closest("button, a");if(a){let i=a;g("click",{tagName:i.tagName.toLowerCase(),id:i.id,className:i.className,text:i.innerText||i.textContent,href:i.href},{batcher:o,visitorId:t,sessionId:n})}},!0)};var y=(o,t={})=>{let n={userAgent:navigator.userAgent,language:navigator.language,screen:`${window.screen.width}x${window.screen.height}`,hostname:window.location.hostname},e={apiKey:o,context:n,...t.batchSize&&{batchSize:t.batchSize},...t.batchTimeout&&{timeout:t.batchTimeout}},a={batcher:l(e),visitorId:p(),sessionId:h()};g("page_view",{},a),window.addEventListener("popstate",()=>g("page_view",{},a)),u(a)};export{y as pulse};
+var l=async o=>{let t="https://pulsed.pages.dev/api/ingest";if(o.events.length===0)return;let e=JSON.stringify(o);if(navigator.sendBeacon){let n=new Blob([e],{type:"application/json"});if(navigator.sendBeacon(t,n))return}try{await fetch(t,{method:"POST",headers:{"Content-Type":"application/json"},body:e,keepalive:!0})}catch(n){console.error("Analytics: Failed to send batch",n)}};var x={batchSize:10,timeout:5e3},p=o=>{let t={...x,...o},e=[],n=null,s=async()=>{if(n&&(clearTimeout(n),n=null),e.length===0)return;let a={apiKey:t.apiKey,context:t.context,events:[...e]};e=[],await l(a)};return{add:a=>{if(e.push({...a}),e.length>=t.batchSize){s();return}n&&clearTimeout(n),n=setTimeout(s,t.timeout)},flush:s}};var u=()=>{let o="analytics_visitor_id",t=localStorage.getItem(o);return t||(t=m(),localStorage.setItem(o,t)),t},h=()=>{let t=sessionStorage.getItem("_sid"),e=sessionStorage.getItem("_sid_t"),n=Date.now();if(t&&e&&n-parseInt(e)<18e5)return sessionStorage.setItem("_sid_t",n.toString()),t;let s=m();return sessionStorage.setItem("_sid",s),sessionStorage.setItem("_sid_t",n.toString()),s},m=()=>Date.now().toString(36)+Math.random().toString(36).substring(2);var g=(o,t={},{batcher:e,visitorId:n,sessionId:s})=>{let i={type:o,visitorId:n,sessionId:s,timestamp:Date.now(),properties:{...t,path:window.location.pathname,query:window.location.search,title:document.title,...o==="page_view"&&{referrer:document.referrer}}};e.add(i)},f=({batcher:o,visitorId:t,sessionId:e})=>{document.addEventListener("click",n=>{let i=n.target.closest("button, a");if(i){let a=i;g("click",{tagName:a.tagName.toLowerCase(),id:a.id,className:a.className,text:a.innerText||a.textContent,href:a.href},{batcher:o,visitorId:t,sessionId:e})}},!0)};var r=Array.from(document.querySelectorAll("script")).find(o=>o.src.includes("pulse.js")),y=r?.dataset.apiKey,v=r?.dataset.batchSize?parseInt(r.dataset.batchSize):void 0,S=r?.dataset.batchTimeout?parseInt(r.dataset.batchTimeout):void 0;if(!y)console.error("[Pulse] Missing data-api-key attribute");else{let o={userAgent:navigator.userAgent,language:navigator.language,screen:`${window.screen.width}x${window.screen.height}`,hostname:window.location.hostname},e={batcher:p({apiKey:y,context:o,...v&&{batchSize:v},...S&&{timeout:S}}),visitorId:u(),sessionId:h()};g("page_view",{},e),window.addEventListener("popstate",()=>g("page_view",{},e)),f(e)}
