+++ apps/script/index.html
@@ -25,11 +25,7 @@
         <button class="nested">Nested Button</button>
     </div>
 
-    <script type="module">
-        import { pulse } from './dist/index.js';
-
-        pulse('API_KEY');
-    </script>
+    <script src="./dist/pulse.js" data-api-key="1234567890" data-batch-size="10" data-batch-timeout="5000"></script>
 </body>
 
 </html>
