+++ apps/script/src/sender.ts
@@ -1,5 +1,5 @@
 export const sendBatch = async (batch: { events: any[], apiKey: string }) => {
-    const endpoint = `https://api.pulsed.workers.dev/ingest`;
+    const endpoint = `https://pulsed.pages.dev/api/ingest`;
     if (batch.events.length === 0) return;
 
     const payload = JSON.stringify(batch);
